Add Zydis dependancy
At least for now, we will be using Zydis for use in injection code that needs to analyze a code section. This is to allow the process to be debugged during injection (by using a jmp instead of int3). We may need to switch to something like Capstone in the future due to Zydis only supporting x86 and x86-64.